convert a string to an integer

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -7,7 +7,7 @@
  */
 int _atoi(char *s)
 {
-	unsigned int sign = 1, num = 0, start = 0;
+	long int sign = 1, num = 0, start = 0;
 
 	while (*s != '\0')
 	{
